requirements.txt
arrow==0.7.0
beautifulsoup4==4.4.1
dateutils==0.6.6
feedparser==5.2.1
Jinja2==2.8
MarkupSafe==0.23
peewee==2.8.0
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
six==1.10.0
Unidecode==0.4.19
wheel==0.24.0
